+++ backend/localrep/models/__init__.py
@@ -3,7 +3,6 @@ from .computeplan import ComputePlan
 from .computetask import ComputeTask
 from .datamanager import DataManager
 from .datasample import DataSample
-from .metric import Metric
 from .model import Model
 from .node import ChannelNode
 from .performance import Performance
@@ -14,7 +13,6 @@ __all__ = [
     "ComputeTask",
     "DataManager",
     "DataSample",
-    "Metric",
     "Model",
     "ChannelNode",
     "Performance",