+++ backend/substrapp/models/__init__.py
@@ -4,12 +4,10 @@ from .computeplan_worker_mapping import ComputePlanWorkerMapping
 from .datamanager import DataManager
 from .datasample import DataSample
 from .image_entrypoint import ImageEntrypoint
-from .metric import Metric
 from .model import Model
 
 __all__ = [
     "DataSample",
-    "Metric",
     "DataManager",
     "Algo",
     "Model",