+++ backend/substrapp/serializers/__init__.py
@@ -1,11 +1,9 @@
 from .algo import AlgoSerializer
 from .datamanager import DataManagerSerializer
 from .datasample import DataSampleSerializer
-from .metric import MetricSerializer
 
 __all__ = [
     "DataSampleSerializer",
     "DataManagerSerializer",
-    "MetricSerializer",
     "AlgoSerializer",
 ]