+++ backend/substrapp/tests/assets.py
@@ -37,11 +37,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:18:53.683675+00:00",
             "description": {
                 "checksum": "a77883e089f18b740805d866bd4170e198815e04320d5810f11c09d4aaec274d",
-                "storage_address": "http://testserver/metric/580d04d4-8dae-45c7-9be9-5beb4b0037d7/description/"
+                "storage_address": "http://testserver/algo/580d04d4-8dae-45c7-9be9-5beb4b0037d7/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/580d04d4-8dae-45c7-9be9-5beb4b0037d7/metrics/"
+                "storage_address": "http://testserver/algo/580d04d4-8dae-45c7-9be9-5beb4b0037d7/file/"
             }
         },
         {
@@ -67,11 +67,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:18:54.247935+00:00",
             "description": {
                 "checksum": "e2acc0c8b9460a75847aa187e90955dd16d57ab07d9738c151def4b5b44aa3ec",
-                "storage_address": "http://testserver/metric/85c572a8-0a23-4fea-8f47-0576df3ecedb/description/"
+                "storage_address": "http://testserver/algo/85c572a8-0a23-4fea-8f47-0576df3ecedb/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/85c572a8-0a23-4fea-8f47-0576df3ecedb/metrics/"
+                "storage_address": "http://testserver/algo/85c572a8-0a23-4fea-8f47-0576df3ecedb/file/"
             }
         },
         {
@@ -97,11 +97,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:18:54.822837+00:00",
             "description": {
                 "checksum": "e95988e3e5b47125fffefa82a898620285b8421af3c5835167bfbd4148c7d760",
-                "storage_address": "http://testserver/metric/b4490c5e-2ae8-420a-9895-55904873d27a/description/"
+                "storage_address": "http://testserver/algo/b4490c5e-2ae8-420a-9895-55904873d27a/description/"
             },
             "algorithm": {
                 "checksum": "6d59e5a01069c10d3bddd271b6a72e90fd5eb1cc2b4b6452e3606f31afd2e9f7",
-                "storage_address": "http://testserver/metric/b4490c5e-2ae8-420a-9895-55904873d27a/metrics/"
+                "storage_address": "http://testserver/algo/b4490c5e-2ae8-420a-9895-55904873d27a/file/"
             }
         },
         {
@@ -127,11 +127,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:18:55.204751+00:00",
             "description": {
                 "checksum": "8245d903d3e29ae3ef3efe12b4558ceab144c1fa920f22cd199585c479d1f203",
-                "storage_address": "http://testserver/metric/1db4c911-732a-4ee2-9fee-2ad08730edc2/description/"
+                "storage_address": "http://testserver/algo/1db4c911-732a-4ee2-9fee-2ad08730edc2/description/"
             },
             "algorithm": {
                 "checksum": "47de8cdf59f84c322f69f9db073fbe9bfbc17f23749b9711916426b65777b75a",
-                "storage_address": "http://testserver/metric/1db4c911-732a-4ee2-9fee-2ad08730edc2/metrics/"
+                "storage_address": "http://testserver/algo/1db4c911-732a-4ee2-9fee-2ad08730edc2/file/"
             }
         },
         {
@@ -157,11 +157,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:18:55.527738+00:00",
             "description": {
                 "checksum": "9782cbb244775b4015b091f83116e2e189694766ab49e34702d8efa412095820",
-                "storage_address": "http://testserver/metric/b6ac7389-bc92-47ae-8c20-856d5d58995d/description/"
+                "storage_address": "http://testserver/algo/b6ac7389-bc92-47ae-8c20-856d5d58995d/description/"
             },
             "algorithm": {
                 "checksum": "c2900cb6bb04e20640fa1e0720365c3c9b509b4d05a0672cd35f118c0b3156a8",
-                "storage_address": "http://testserver/metric/b6ac7389-bc92-47ae-8c20-856d5d58995d/metrics/"
+                "storage_address": "http://testserver/algo/b6ac7389-bc92-47ae-8c20-856d5d58995d/file/"
             }
         },
         {
@@ -187,11 +187,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:18:57.145250+00:00",
             "description": {
                 "checksum": "f1ae963cc9a18ae84d220912299c45842c874bacd4f2b0a5f523c9bcc3c9d25e",
-                "storage_address": "http://testserver/metric/79a749ec-32fe-483f-a3fe-4385c6467842/description/"
+                "storage_address": "http://testserver/algo/79a749ec-32fe-483f-a3fe-4385c6467842/description/"
             },
             "algorithm": {
                 "checksum": "47de8cdf59f84c322f69f9db073fbe9bfbc17f23749b9711916426b65777b75a",
-                "storage_address": "http://testserver/metric/79a749ec-32fe-483f-a3fe-4385c6467842/metrics/"
+                "storage_address": "http://testserver/algo/79a749ec-32fe-483f-a3fe-4385c6467842/file/"
             }
         },
         {
@@ -217,11 +217,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:19:12.478571+00:00",
             "description": {
                 "checksum": "0408c739620d9e5643d9a71a4a886d6344fc7345af776f75005758adc2459e4d",
-                "storage_address": "http://testserver/metric/95869305-106e-4fbd-9c58-8889df4d64c9/description/"
+                "storage_address": "http://testserver/algo/95869305-106e-4fbd-9c58-8889df4d64c9/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/95869305-106e-4fbd-9c58-8889df4d64c9/metrics/"
+                "storage_address": "http://testserver/algo/95869305-106e-4fbd-9c58-8889df4d64c9/file/"
             }
         },
         {
@@ -247,11 +247,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:19:13.034036+00:00",
             "description": {
                 "checksum": "2da5968b0f3ca9185b4a7c6367767718fedc0a8bf5f294b4cdf6ad0d442d21bd",
-                "storage_address": "http://testserver/metric/13b1aa4b-7967-41b5-94fe-7d2faa3b8a8c/description/"
+                "storage_address": "http://testserver/algo/13b1aa4b-7967-41b5-94fe-7d2faa3b8a8c/description/"
             },
             "algorithm": {
                 "checksum": "47de8cdf59f84c322f69f9db073fbe9bfbc17f23749b9711916426b65777b75a",
-                "storage_address": "http://testserver/metric/13b1aa4b-7967-41b5-94fe-7d2faa3b8a8c/metrics/"
+                "storage_address": "http://testserver/algo/13b1aa4b-7967-41b5-94fe-7d2faa3b8a8c/file/"
             }
         },
         {
@@ -277,11 +277,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:19:45.696136+00:00",
             "description": {
                 "checksum": "08e24e439d68ab5f82da695600c798c9cdde57454731f3871692140f54fa0fbb",
-                "storage_address": "http://testserver/metric/6a9750f5-76d2-4e6b-839f-eabc0611427c/description/"
+                "storage_address": "http://testserver/algo/6a9750f5-76d2-4e6b-839f-eabc0611427c/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/6a9750f5-76d2-4e6b-839f-eabc0611427c/metrics/"
+                "storage_address": "http://testserver/algo/6a9750f5-76d2-4e6b-839f-eabc0611427c/file/"
             }
         },
         {
@@ -307,11 +307,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:19:48.577301+00:00",
             "description": {
                 "checksum": "d9b87dc7f27075c16264b76286cff509bb028432ca80e7e260b227e26db8b1f1",
-                "storage_address": "http://testserver/metric/aa86e900-9268-481b-acdd-07dfe01f21d3/description/"
+                "storage_address": "http://testserver/algo/aa86e900-9268-481b-acdd-07dfe01f21d3/description/"
             },
             "algorithm": {
                 "checksum": "47de8cdf59f84c322f69f9db073fbe9bfbc17f23749b9711916426b65777b75a",
-                "storage_address": "http://testserver/metric/aa86e900-9268-481b-acdd-07dfe01f21d3/metrics/"
+                "storage_address": "http://testserver/algo/aa86e900-9268-481b-acdd-07dfe01f21d3/file/"
             }
         },
         {
@@ -337,11 +337,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:20:08.392073+00:00",
             "description": {
                 "checksum": "958d91b5511921bdca18934bb7c706d0dbd4ac794307dd7e49f1ed61ac8cb1d7",
-                "storage_address": "http://testserver/metric/e3986b54-27d7-4e15-ad1f-30f08fc414ee/description/"
+                "storage_address": "http://testserver/algo/e3986b54-27d7-4e15-ad1f-30f08fc414ee/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/e3986b54-27d7-4e15-ad1f-30f08fc414ee/metrics/"
+                "storage_address": "http://testserver/algo/e3986b54-27d7-4e15-ad1f-30f08fc414ee/file/"
             }
         },
         {
@@ -367,11 +367,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:20:13.537647+00:00",
             "description": {
                 "checksum": "65e4858123af91f5693479853389d44df5ca6a66e595a2af7fc62b1d9c4139d1",
-                "storage_address": "http://testserver/metric/024cda63-7492-47fa-9dc4-1b6c7062f9d1/description/"
+                "storage_address": "http://testserver/algo/024cda63-7492-47fa-9dc4-1b6c7062f9d1/description/"
             },
             "algorithm": {
                 "checksum": "47de8cdf59f84c322f69f9db073fbe9bfbc17f23749b9711916426b65777b75a",
-                "storage_address": "http://testserver/metric/024cda63-7492-47fa-9dc4-1b6c7062f9d1/metrics/"
+                "storage_address": "http://testserver/algo/024cda63-7492-47fa-9dc4-1b6c7062f9d1/file/"
             }
         },
         {
@@ -397,19 +397,19 @@ def get_metrics():
             "creation_date": "2022-01-20T14:24:19.918194+00:00",
             "description": {
                 "checksum": "27517ac61a82dae6b373a20224d09c444119cd5a7b4a5f291f22dbdb6b743113",
-                "storage_address": "http://testserver/metric/c7f1c5d1-19fa-46d7-b381-f5f1a7b1645f/description/"
+                "storage_address": "http://testserver/algo/c7f1c5d1-19fa-46d7-b381-f5f1a7b1645f/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/c7f1c5d1-19fa-46d7-b381-f5f1a7b1645f/metrics/"
+                "storage_address": "http://testserver/algo/c7f1c5d1-19fa-46d7-b381-f5f1a7b1645f/file/"
             }
         },
         {
             "key": "df82fb26-1397-4883-b54b-7e2c7d691071",
             "name": "13d5042df7f24ecfa01ca8e9bcc8e822_test_asset_with_metadata[metadata0-metadat...13d5042df7f24ecfa01c",
             "owner": "MyOrg1MSP",
+            "category": "ALGO_METRIC",
             "metadata": {
-                "category": "ALGO_METRIC",
                 "foo": "bar"
             },
             "permissions": {
@@ -429,11 +429,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:24:20.358876+00:00",
             "description": {
                 "checksum": "d6dfecefa762664a312334f48f9b9d92b786a09100231ef87ceac61c94f0aa0f",
-                "storage_address": "http://testserver/metric/df82fb26-1397-4883-b54b-7e2c7d691071/description/"
+                "storage_address": "http://testserver/algo/df82fb26-1397-4883-b54b-7e2c7d691071/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/df82fb26-1397-4883-b54b-7e2c7d691071/metrics/"
+                "storage_address": "http://testserver/algo/df82fb26-1397-4883-b54b-7e2c7d691071/file/"
             }
         },
         {
@@ -459,11 +459,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:24:20.927074+00:00",
             "description": {
                 "checksum": "762c9219987930177703bc71e9b8c1ffcdb1a3ceab3a8f34bf050831667d39b1",
-                "storage_address": "http://testserver/metric/f3d26d7e-e957-4412-9b1c-581c13aa189f/description/"
+                "storage_address": "http://testserver/algo/f3d26d7e-e957-4412-9b1c-581c13aa189f/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/f3d26d7e-e957-4412-9b1c-581c13aa189f/metrics/"
+                "storage_address": "http://testserver/algo/f3d26d7e-e957-4412-9b1c-581c13aa189f/file/"
             }
         },
         {
@@ -489,11 +489,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:24:21.784010+00:00",
             "description": {
                 "checksum": "fcdc71ed32c412b9cc4d95bae063c1bddd72ba01b34681ee76cd0991531408d4",
-                "storage_address": "http://testserver/metric/6a7bcbb4-78f3-4d48-8b7d-2851983cf8f1/description/"
+                "storage_address": "http://testserver/algo/6a7bcbb4-78f3-4d48-8b7d-2851983cf8f1/description/"
             },
             "algorithm": {
                 "checksum": "2873791633e70cb822a46c68a9611df34889fcf1cf317bdc05907e107c6f6c4f",
-                "storage_address": "http://testserver/metric/6a7bcbb4-78f3-4d48-8b7d-2851983cf8f1/metrics/"
+                "storage_address": "http://testserver/algo/6a7bcbb4-78f3-4d48-8b7d-2851983cf8f1/file/"
             }
         },
         {
@@ -519,11 +519,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:26:19.633416+00:00",
             "description": {
                 "checksum": "77ef974020b39954f873a8cac8c938e3ecba3f8e1049e168ee5ab60ca47c2b43",
-                "storage_address": "http://testserver/metric/1710cd9c-d34f-4fbf-822e-58be9c0a92b2/description/"
+                "storage_address": "http://testserver/algo/1710cd9c-d34f-4fbf-822e-58be9c0a92b2/description/"
             },
             "algorithm": {
                 "checksum": "ab2a45a85b2c25cb09ff85ce6899b826049e1da7cee31ae2033f5cabc2c37f11",
-                "storage_address": "http://testserver/metric/1710cd9c-d34f-4fbf-822e-58be9c0a92b2/metrics/"
+                "storage_address": "http://testserver/algo/1710cd9c-d34f-4fbf-822e-58be9c0a92b2/file/"
             }
         },
         {
@@ -549,11 +549,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:26:26.945760+00:00",
             "description": {
                 "checksum": "13a57263d9c7eb41c83dc754d21ced689838dd0d5f4c5ec9a6f41ed91faa2a6f",
-                "storage_address": "http://testserver/metric/3fd33f3e-1071-4ac9-9551-0cca91695e77/description/"
+                "storage_address": "http://testserver/algo/3fd33f3e-1071-4ac9-9551-0cca91695e77/description/"
             },
             "algorithm": {
                 "checksum": "ab2a45a85b2c25cb09ff85ce6899b826049e1da7cee31ae2033f5cabc2c37f11",
-                "storage_address": "http://testserver/metric/3fd33f3e-1071-4ac9-9551-0cca91695e77/metrics/"
+                "storage_address": "http://testserver/algo/3fd33f3e-1071-4ac9-9551-0cca91695e77/file/"
             }
         },
         {
@@ -579,11 +579,11 @@ def get_metrics():
             "creation_date": "2022-01-20T14:26:27.139952+00:00",
             "description": {
                 "checksum": "1a3270a36637653029262671acd63bf645479011bac7680930e9a98375a20d80",
-                "storage_address": "http://testserver/metric/2d406537-6893-47af-99ac-be507a25413a/description/"
+                "storage_address": "http://testserver/algo/2d406537-6893-47af-99ac-be507a25413a/description/"
             },
             "algorithm": {
                 "checksum": "57e5fa23058671aa57df3e694bf67cfadd527ff16a1ab2cb1aebb3a3fc2c992d",
-                "storage_address": "http://testserver/metric/2d406537-6893-47af-99ac-be507a25413a/metrics/"
+                "storage_address": "http://testserver/algo/2d406537-6893-47af-99ac-be507a25413a/file/"
             }
         }
     ]